Create Exercícios B 1 - Use Lists in Kotlin
It's common to make lists for all sorts of situations in your everyday life such as a list of things to do, a list of guests for an event, a wish list, or a grocery list. In programming, lists are also very useful. For example, there could be a list of news articles, songs, calendar events, or social media posts within an app.

Learning how to create and use lists is an important programming concept to add to your toolbox, and it will enable you to create more sophisticated apps.

In this codelab, you will use the Kotlin Playground to become familiar with lists in Kotlin and create a program for ordering different variations of noodle soup. Are you hungry yet?


Signed-off-by: RitaPintoBlissApp <144903905+RitaPintoBlissApp@users.noreply.github.com>